# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FitAI API",
    "version": "1.0.0",
    "description": "API oficial do FitAI - Inteligência Artificial para Fitness",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Suporte FitAI",
      "url": "https://fitai.com/support",
      "email": "support@fitai.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/exercises": {
      "get": {
        "summary": "Busca exercícios com filtros",
        "description": "Retorna uma lista de exercícios permitindo filtrar por nome, grupo muscular ou equipamento.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "example": "bench press"
          },
          {
            "in": "query",
            "name": "bodyPart",
            "schema": {
              "type": "string"
            },
            "example": "chest"
          },
          {
            "in": "query",
            "name": "equipment",
            "schema": {
              "type": "string"
            },
            "example": "barbell"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de exercícios retornada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "0001"
                          },
                          "name": {
                            "type": "string",
                            "example": "Bench Press"
                          },
                          "body_part": {
                            "type": "string",
                            "example": "chest"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{id}": {
      "get": {
        "summary": "Detalhes de um exercício",
        "description": "Retorna informações completas de um exercício pelo ID.",
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0001"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}